Fixed 2 bugs related to bootloader reset/entering sequence
-Fixed: weired bug where input pin BOOT1 when set to pull up/down it will always stay HIGH/LOW, depending on the active pull up/down resistor.
the solution is using a floating mode

-Fixed: GPIO_readPin() API, it must take a u32 out pointer rather than u8, since it returns the pin value not-shifted

+++ include/GPIO_interface.h
@@ -109,6 +109,6 @@ extern ERROR_STATUS GPIO_initPin(GPIO_t * peri);
 extern ERROR_STATUS GPIO_writePin(GPIO_t * peri, u32 value);
 
 
-extern ERROR_STATUS GPIO_readPin(GPIO_t * peri, u8 *value);
+extern ERROR_STATUS GPIO_readPin(GPIO_t * peri, u32 *value);
 
 #endif
